mgr/dashboard: add password expiry date
Add a 'pwd_expiry_date' field to the User object to be able
to set password expiry date per user.

There are two options to set a password expiry date for a
user:

1. Set the 'USER_PWD_DEFAULT_EXPIRY_SPAN' setting. If defined
   the expiry date of the password will be calculated
   automatically based on the given value. It will also be
   updated automatically when the user changes the password.
2. Set an expiry date by setting the 'pwdexpirydate' field
   when creating a user.

Add two more settings in addition: USER_PWD_EXPIRY_WARNING_1
and USER_PWD_EXPIRY_WARNING_2. These two settings are defining
the amount of days to notify the user that his password will
expiry soon.

It's not possible to set the 'pwd_expiry_date' field to a
date in the past.

If the password of a user is already expired he is no longer
allowed to log into the dashboard.

Fixes: https://tracker.ceph.com/issues/40329
Signed-off-by: Tatjana Dehler <tdehler@suse.com>