mgr/dashboard: comment questionable suites out
We're currently facing some issues with our integration
tests. Because of that we agreed on commenting questionable
suites out to be able to run all other suites on open pull
requests.

'test_health' and 'test_perf_counters' are commented out
because they led to issues in relation to
https://tracker.ceph.com/issues/41538
As soon as the issue has been fixed, we need to re-add
these two suites again.

'test_orchestrator' and 'test_rbd_mirroring' are commented
out because they are pretty new and have never been running
before. We can integrate them later as soon as the tests
are more stable again.

Signed-off-by: Tatjana Dehler <tdehler@suse.com>