mgr/cephadm: allow unsigned plugins
From Grafana v8.0 all plugins are required to be signed:
https://grafana.com/docs/grafana/latest/installation/upgrading/#plugins

We're using plugins whose highest available version (on Github) is
unsigned. As a result the plugin within the Grafana dashboard can't
be loaded.
We can workaround it for now by adding the
`allow_loading_unsigned_plugins`option to the grafana.ini template.
Once all required plugins are signed we should remove the option again.

Signed-off-by: Tatjana Dehler <tdehler@suse.com>